Update Local Alignment.py
Without the extra break, you will still be in the nested for-loop and this causes the program to actually consider the wrong highest score. 

if you for example set the:
v = "PAWHEAE"
w = "HDAGAWGHEQ"

the result is:
AW-HEA
AWGHEQ 

while it should be 
AW-HE
AWGHE 

What do you think? I hope this is right.